+++ panther/panther_hardware.repos
@@ -11,6 +11,10 @@ repositories:
     type: git
     url: https://github.com/husarion/panther_msgs.git
     version: ros2
+  robot_localization: # Caused by this error https://github.com/cra-ros-pkg/robot_localization/pull/876, it can be removed after the next synchronization of apt packages.
+    type: git
+    url: https://github.com/cra-ros-pkg/robot_localization.git
+    version: humble-devel
   ros_components_description:
     type: git
     url: https://github.com/husarion/ros_components_description.git
